Tests: Factor LoggerMock out in own class.
